+++ src/cfn-mcp-server/CHANGELOG.md
@@ -16,4 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Initial project setup
+- Support for CloudFormation Template generation via IaC Generator APIs
+
+## [0.0.1] 2025-05-14
+
+### Added
+
+- Initial release of CloudFormation MCP Server
+- Support for resource CRUDL via CloudControl APIs